+++ src/leiningen/new/chestnut/.dir-locals.el
@@ -1,3 +1,3 @@
 ((nil . ((cider-refresh-before-fn . "reloaded.repl/suspend")
          (cider-refresh-after-fn  . "reloaded.repl/resume")
-         (cider-cljs-lein-repl    . "(do (user/run) (user/browser-repl))"))))
+         (cider-cljs-lein-repl    . "(do (user/go) (user/cljs-repl))"))))